requirements.txt
asgiref==3.6.0
boto3==1.28.20
botocore==1.31.20
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.2.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.16
